Switch back to double with QuarticSolver, as profiling seems to say its faster...
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@197433 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 38a41199339a9368546fce79cb1d5fc24f0db175

+++ Rec/Rich/RichRecPhotonTools/src/application/PhotonReco/main.cpp
@@ -43,11 +43,11 @@ public:
 
 bool solve( const Data& data )
 {
-  return qSolver.solve<float>( data.emissPnt, 
-                               data.centOfCurv, 
-                               data.virtDetPoint,
-                               data.radius, 
-                               sphReflPoint );
+  return qSolver.solve<double>( data.emissPnt, 
+                                data.centOfCurv, 
+                                data.virtDetPoint,
+                                data.radius, 
+                                sphReflPoint );
 }
 
 void solve( const Data::Vector & dataV )
